rename: 채팅 아이콘 파일 위치 변경
